requirements.txt
autogen-core==0.7.2
autogen-agentchat==0.7.2
autogen-ext==0.7.2
autogen-ext[openai]
python-dotenv==1.1.1
ipykernel==6.30.1
pydantic==2.11.7
pydantic[email]
PyPDF2==3.0.1
fastapi==0.116.1
uvicorn==0.35.0
python-multipart==0.0.20
pymupdf==1.26.3
python-docx==1.2.0
SQLAlchemy==2.0.43
psycopg2-binary==2.9.10